requirements.txt
dotenv
python-telegram-bot==21.0.1 
apscheduler 
python-telegram-bot-calendar
aiogram
aiohttp
Pillow